// src/types.ts
export interface Community {
    id: string;
    fillColor: string;
    borderColor: string;
    outline: POINT[];
}
// ...
export interface GraphNode extends vis.Node {
}

export interface LoadGraphOption {
    dynamic: boolean;
    centerPointX?: number;
    centerPointY?: number;
    scale?: number;
}

export interface LoadGraphOptionCallback {
    width: number;
    height: number;
}

export interface NodeCommunityEntry {
    node: string;
    community: string;
}

export interface InitData {
    nodesNum: number;
    edgesNum: number;
    autoLayout: boolean;
}

export interface CommunityData {
    communities: Community[];
    nodeMap: NodeCommunityEntry[];
}

export interface GraphEdge extends vis.Edge {

}
// ...
export interface RELATION_PATH {
    nodes: object[];
    edges: object[];
}

export interface QUERY_RESULTS {
    completed: boolean;
    paths: RELATION_PATH[];
}

export interface PAIR<K, V> {
    _1: K;
    _2: V;
}

export interface RECT {
    left: number;
    top: number;
    right: number;
    bottom: number;
}

export interface POINT {
    x: number;
    y: number;
}